Move all socket creation stuff to transport code
